requirements.txt
flask==3.0.3
pillow==10.2.0

# çekirdek bilimsel kütüphaneler
numpy==1.26.4

# YOLO + PyTorch
ultralytics==8.3.0
torch==2.3.1
torchvision==0.18.1

# istek atmak için (gerekirse)
requests==2.32.3
